First Pass at Packaging RequireJS Dependencies
