+++ .travis.yml
@@ -10,7 +10,8 @@ before_install:
 install:
     - make test.acceptance
     - pip install coveralls
-script: 
+script:
+    - make static
     - make validate
     - bash ./runAcceptance.sh
 after_success:
