+++ analytics_dashboard/courses/templates/courses/engagement_content.html
@@ -13,7 +13,7 @@ Individual course-centric engagement content view.
 
 {% block javascript %}
   {{ block.super }}
-  <script src="{% static 'js/engagement-content-main.js' %}"></script>
+  <script src="{% static 'dist/js/engagement-content-main.js' %}"></script>
 {% endblock javascript %}
 
 {% block child_content %}
