+++ analytics_dashboard/courses/templates/courses/enrollment_activity.html
@@ -11,7 +11,7 @@ Individual course-centric enrollment activity view.
 
 {% block javascript %}
   {{ block.super }}
-  <script src="{% static 'js/enrollment-activity-main.js' %}"></script>
+  <script src="{% static 'dist/js/enrollment-activity-main.js' %}"></script>
 {% endblock javascript %}
