+++ analytics_dashboard/courses/templates/courses/enrollment_geography.html
@@ -11,7 +11,7 @@ Individual course-centric enrollment geography view.
 
 {% block javascript %}
   {{ block.super }}
-  <script src="{% static 'js/enrollment-geography-main.js' %}"></script>
+  <script src="{% static 'dist/js/enrollment-geography-main.js' %}"></script>
 {% endblock javascript %}
 
 {% block child_content %}
