--- analytics_dashboard/static/js/application-main.js
@@ -1,7 +1,0 @@
-/**
- * Load scripts needed across the application.
- */
-
-require(['bootstrap', 'bootstrap_accessibility', 'vendor/domReady!', 'load/init-page'], function (bootstrap, bootstrap_accessibility, doc, page) {
-    'use strict';
-});
