+++ analytics_dashboard/static/js/test/spec-runner.js
@@ -5,9 +5,7 @@
 
 var isBrowser = window.__karma__ === undefined;
 var specs = [];
-
-// Loaded from js/common.js
-var config = require_config;
+var config = {};
 
 // Two execution paths: browser or gulp
 if(isBrowser) {
