+++ karma.conf.js
@@ -22,7 +22,7 @@ module.exports = function (config) {
             {pattern: 'analytics_dashboard/static/js/views/**/*.js', included: false},
             {pattern: 'analytics_dashboard/static/js/utils/**/*.js', included: false},
             {pattern: 'analytics_dashboard/static/js/test/specs/*.js', included: false},
-            'analytics_dashboard/static/js/common.js',
+            'analytics_dashboard/static/js/config.js',
             'analytics_dashboard/static/js/test/spec-runner.js'
         ],
