+++ package.json
@@ -10,7 +10,7 @@
   },
   "devDependencies": {
     "browser-sync": "^1.1.2",
-    "gulp": "^3.8.1",
+    "gulp": "^3.8.8",
     "gulp-jshint": "^1.6.3",
     "gulp-karma": "0.0.4",
     "karma": "^0.12.16",
